+++ Marlin/src/lcd/menu/menu_configuration.cpp
@@ -35,7 +35,7 @@
 #endif
 
 #if HAS_FANCHECK
-  #include "../../module/fancheck.h"
+  #include "../../feature/fancheck.h"
 #endif
 
 #if ENABLED(POWER_LOSS_RECOVERY)
