+++ ini/features.ini
@@ -115,6 +115,7 @@ I2C_POSITION_ENCODERS                  = src_filter=+<src/feature/encoder_i2c.cp
 IIC_BL24CXX_EEPROM                     = src_filter=+<src/libs/BL24CXX.cpp>
 HAS_SPI_FLASH                          = src_filter=+<src/libs/W25Qxx.cpp>
 HAS_ETHERNET                           = src_filter=+<src/feature/ethernet.cpp> +<src/gcode/feature/network/M552-M554.cpp>
+HAS_FANCHECK                           = src_filter=+<src/feature/fancheck.cpp> +<src/gcode/temp/M123.cpp>
 HAS_FANMUX                             = src_filter=+<src/feature/fanmux.cpp>
 FILAMENT_WIDTH_SENSOR                  = src_filter=+<src/feature/filwidth.cpp> +<src/gcode/feature/filwidth>
 FWRETRACT                              = src_filter=+<src/feature/fwretract.cpp> +<src/gcode/feature/fwretract>
@@ -210,7 +211,6 @@ SDSUPPORT                              = src_filter=+<src/sd/cardreader.cpp> +<s
 HAS_MEDIA_SUBCALLS                     = src_filter=+<src/gcode/sd/M32.cpp>
 GCODE_REPEAT_MARKERS                   = src_filter=+<src/feature/repeat.cpp> +<src/gcode/sd/M808.cpp>
 HAS_EXTRUDERS                          = src_filter=+<src/gcode/units/M82_M83.cpp> +<src/gcode/temp/M104_M109.cpp> +<src/gcode/config/M221.cpp>
-HAS_FANCHECK                           = src_filter=+<src/gcode/temp/M123.cpp> +<src/module/fancheck.cpp>
 HAS_TEMP_PROBE                         = src_filter=+<src/gcode/temp/M192.cpp>
 HAS_COOLER                             = src_filter=+<src/gcode/temp/M143_M193.cpp>
 AUTO_REPORT_TEMPERATURES               = src_filter=+<src/gcode/temp/M155.cpp>
