+++ platformio.ini
@@ -111,6 +111,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/feature/e_parser.cpp>
   -<src/feature/encoder_i2c.cpp>
   -<src/feature/ethernet.cpp> -<src/gcode/feature/network/M552-M554.cpp>
+  -<src/feature/fancheck.cpp>
   -<src/feature/fanmux.cpp>
   -<src/feature/filwidth.cpp> -<src/gcode/feature/filwidth>
   -<src/feature/fwretract.cpp> -<src/gcode/feature/fwretract>
@@ -241,7 +242,6 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/libs/least_squares_fit.cpp>
   -<src/libs/nozzle.cpp> -<src/gcode/feature/clean>
   -<src/module/delta.cpp>
-  -<src/module/fancheck.cpp>
   -<src/module/planner_bezier.cpp>
   -<src/module/polargraph.cpp>
   -<src/module/printcounter.cpp>
